--- src/Algorithm/Arrays/Best Time to Buy and Sell Stock/bestTimeToBuyStock.java
@@ -1,14 +1,0 @@
-// Problem Link : https://leetcode.com/problems/best-time-to-buy-and-sell-stock/
-
-class Solution {
-
-    public int maxProfit(int[] prices) {
-        int min = Integer.MAX_VALUE;
-        int ans = Integer.MIN_VALUE;
-        for (int val : prices) {
-            min = Math.min(min, val);
-            ans = Math.max(ans, val - min);
-        }
-        return ans;
-    }
-}
