--- src/Algorithm/Arrays/Best Time to Buy and Sell Stock/bestTimeToBuyStock.py
@@ -1,12 +1,0 @@
-# Problem Link : https://leetcode.com/problems/best-time-to-buy-and-sell-stock/
-
-class Solution:
-    def maxProfit(self, prices: List[int]) -> int:
-        res = 0
-
-        l = 0
-        for r in range(1, len(prices)):
-            if prices[r] < prices[l]:
-                l = r
-            res = max(res, prices[r] - prices[l])
-        return res
