--- src/Algorithm/Arrays/GroupAnagrams/groupAnagrams.java
@@ -1,24 +1,0 @@
-/*
-    Problem link- https://leetcode.com/problems/group-anagrams/
-*/
-*/
-
-class Solution {
-
-    public List<List<String>> groupAnagrams(String[] strs) {
-        List<List<String>> res = new ArrayList<>();
-        if (strs.length == 0) return res;
-        HashMap<String, List<String>> map = new HashMap<>();
-        for (String s : strs) {
-            char[] hash = new char[26];
-            for (char c : s.toCharArray()) {
-                hash[c - 'a']++;
-            }
-            String key = new String(hash);
-            map.computeIfAbsent(key, k -> new ArrayList<>());
-            map.get(key).add(s);
-        }
-        res.addAll(map.values());
-        return res;
-    }
-}
