--- src/Algorithm/Arrays/GroupAnagrams/groupAnagrams.py
@@ -1,16 +1,0 @@
-# Problem link- https://leetcode.com/problems/group-anagrams/
-
-from collections import defaultdict
-from typing import List
-import collections
-
-class Solution:
-    def groupAnagrams(self, strs: List[str]) -> List[List[str]]:
-        ans = collections.defaultdict(list)
-
-        for s in strs:
-            count = [0] * 26
-            for c in s:
-                count[ord(c) - ord("a")] += 1
-            ans[tuple(count)].append(s)
-        return ans.values()
