--- src/Algorithm/Arrays/Longest Consecutive Sequence/lcs.py
@@ -1,21 +1,0 @@
-# Problem Link: https://leetcode.com/problems/longest-consecutive-sequence/
-# Time Complexity: O(n)
-# Space Complexity: O(n)
-
-class Solution:
-    def longestConsecutive(self, nums):
-        longest_streak = 0
-        num_set = set(nums)
-
-        for num in num_set:
-            if num - 1 not in num_set:
-                current_num = num
-                current_streak = 1
-
-                while current_num + 1 in num_set:
-                    current_num += 1
-                    current_streak += 1
-
-                longest_streak = max(longest_streak, current_streak)
-
-        return longest_streak
